# Improve performance of contributors page by a bazillion percent.