--- a/app/commands/user/reputation_token/calculate_contextual_data.rb
+++ b/app/commands/user/reputation_token/calculate_contextual_data.rb
@@ -4,6 +4,14 @@ class User::ReputationToken
     include ActionView::Helpers::TextHelper
     include ActionView::Helpers::NumberHelper
 
+    # This does n queries (where n is number of users).
+    # This is theoretically terrible but actually works out to be much more performant for two reasons:
+    # 1. MySQL's index performance ranges from 0.05s to 33s (!!) depending on how dispersed the indexes pages
+    # are. First page load therefore takes 33s and subsequent ones are pretty instant. 
+    # By reducing it down to n queries, we have a steady performance that's at the very low
+    # end of the range (~0.5s sum for all the queries)
+    # 2. We cache the values per user and invalidate the cache when a new reputation token is
+    # added, so although this is n+1 in the worst case - we're never really actually there.
     def initialize(user_ids, period: nil, earned_since: nil, track_id: nil, category: nil)
       @single_user = user_ids.is_a?(Integer)
       @user_ids = Array(user_ids)
@@ -30,10 +38,8 @@ class User::ReputationToken
 
     memoize
     def data
-      user_ids.each_with_object({}) do |user_id, data|
-        user_tuples = tuples.select { |t| t['user_id'] == user_id }.
-          index_by { |t| t['type'] }.
-          transform_keys { |k| k.split("::").last }
+      tuples.each.with_object({}) do |(user_id, user_tuples), data|
+        user_tuples = user_tuples.index_by { |t| t['type'] }.transform_keys { |k| k.split("::").last }
 
         code_contributions = user_tuples.dig("CodeContributionToken", 'num')
         code_reviews = user_tuples.dig("CodeReviewToken", 'num')
@@ -66,13 +72,19 @@ class User::ReputationToken
 
     memoize
     def tuples
-      ts = User::ReputationToken.where(user_id: user_ids)
-      ts = ts.where(category:) if category
-      ts = ts.where('earned_on >= ?', earned_since) if earned_since
-      ts = ts.where(track_id:) if track_id
-      ts = ts.group(:user_id, :type).select("user_id, type, COUNT(*) as num, SUM(value) as total")
-
-      ActiveRecord::Base.connection.select_all(ts.to_sql)
+      base = User::ReputationToken
+      base = base.where(category:) if category
+      base = base.where('earned_on >= ?', earned_since) if earned_since
+      base = base.where(track_id:) if track_id
+      base = base.group(:type)
+      base = base.select('type, COUNT(value) AS num, SUM(value) AS total')
+ 
+      user_ids.each_with_object({}) do |user_id, hash|
+        hash[user_id] = with_cache(user_id) do
+          query = base.where(user_id: user_id)
+          ActiveRecord::Base.connection.select_all(query)
+        end
+      end
     end
 
     private
@@ -80,5 +92,17 @@ class User::ReputationToken
 
     Data = Struct.new(:activity, :reputation)
     private_constant :Data
+
+    def with_cache(user)
+      user_key = User::ReputationToken.cache_hash_for(user)
+      value_key = ["contextual", earned_since, track_id, category].join("|")
+      redis = Exercism.redis_tooling_client
+      val = redis.hget(user_key, value_key)
+      return JSON.parse(val) if val
+
+      yield.tap do |val|
+        redis.hset(user_key, value_key, val.to_json)
+      end
+    end
   end
 end