--- a/app/commands/user/reputation_token/calculate_contextual_data.rb
+++ b/app/commands/user/reputation_token/calculate_contextual_data.rb
@@ -89,7 +89,6 @@ class User::ReputationToken
 
     private
     attr_reader :user_ids, :single_user, :earned_since, :track_id, :category
-    
 
     Data = Struct.new(:activity, :reputation)
     private_constant :Data