--- a/test/commands/user/reputation_token/calculate_contextual_data_test.rb
+++ b/test/commands/user/reputation_token/calculate_contextual_data_test.rb
@@ -153,5 +153,4 @@ class User::ReputationToken::CalculateContextualDataTest < ActiveSupport::TestCa
       assert_equal 24, data.reputation
     end
   end
-
 end